Refactor auth module
Separate parts
Add user repository
Add user schema
